requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# AI / LangChain
google-generativeai==0.7.2
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.43
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4

# Database
supabase==2.10.0
httpx==0.27.2
websockets==13.1

# Document Processing
python-docx==1.1.2
openpyxl==3.1.5

# Utilities
tenacity==8.5.0
pinecone
